+++ mpf/iset_si.c
@@ -29,7 +29,7 @@ mpf_init_set_si (mpf_ptr r, long int val)
   mp_limb_t vl;
 
   r->_mp_prec = prec;
-  r->_mp_d = (mp_ptr) (*__gmp_allocate_func) ((prec + 1) * BYTES_PER_MP_LIMB);
+  r->_mp_d = (mp_ptr) (*__gmp_allocate_func) ((size_t) (prec + 1) * BYTES_PER_MP_LIMB);
 
   vl = (mp_limb_t) ABS_CAST (unsigned long int, val);
